fixed the coordinates calculation (beneficial for mag tool)

+++ Mouse.h
@@ -3,9 +3,9 @@
 
 class LeftMouse {
 public:
-    sf::Vector2i startMousePos; // start position when mouse was pressed
-    sf::Vector2i lastMousePos; // last known position of mouse
-    sf::Vector2i mousePos;
+    sf::Vector2f startMousePos; // start position when mouse was pressed
+    sf::Vector2f lastMousePos; // last known position of mouse
+    sf::Vector2f mousePos;
 
     LeftMouse();
     void Update(sf::RenderWindow* window);
