+++ Stamps.cpp
@@ -17,13 +17,13 @@ Point::Point():Stamp(0){
     radius = 0;
 }
 
-void Point::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void Point::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     canvas->setPixel(pos.x,pos.y,color);
 }
 
 Circle::Circle(int radius):Stamp(radius){}
 
-void Circle::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void Circle::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     for(int i = -radius; i <= radius; i++){
         for(int j = -radius; j <= radius; j++){
             if(i*i + j*j <= radius*radius){
@@ -37,7 +37,7 @@ Spray::Spray(int radius):Stamp(radius){
     srand(time(0));
 }
 
-void Spray::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void Spray::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     for(int i = -radius; i <= radius; i++){
         for(int j = -radius; j <= radius; j++){
             if((i*i + j*j <= radius*radius) && rand()%100 < 1){
@@ -50,7 +50,7 @@ void Spray::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
 
 Square::Square(int radius):Stamp(radius){}
 
-void Square::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void Square::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     for(int i = -radius; i <= radius; i++){
         for(int j = -radius; j <= radius; j++){
             canvas->setPixel(pos.x+i,pos.y+j,color);
@@ -60,7 +60,7 @@ void Square::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
 
 BackSlash::BackSlash(int radius):Stamp(radius){}
 
-void BackSlash::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void BackSlash::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     for(int i = -radius; i <= radius; i++){
         canvas->setPixel(pos.x+i,pos.y+i,color);       
         if (i+1 <= radius){
@@ -71,7 +71,7 @@ void BackSlash::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
 
 ForwardSlash::ForwardSlash(int radius):Stamp(radius){}
 
-void ForwardSlash::draw(sf::Vector2i pos,sf::Color color,Canvas* canvas){
+void ForwardSlash::draw(sf::Vector2f pos,sf::Color color,Canvas* canvas){
     for(int i = -radius; i <= radius; i++){
         canvas->setPixel(pos.x+i,pos.y-i,color);   
         if (i+1 <= radius){
